Disabled Texture Unit Lookup
When a disabled texture unit is used the last looked up value is
retained.

This code selects the middle pixel of kitten.bin, draws a triangle, then
disables the texture unit, then draws another triangle (with different
texture coordinates).

On actual hardware the second triangle takes the last looked up color of
the first triangle (in this case its the middle pixel of kitten.bin).